+++ engine/runtime/src/main/scala/org/enso/compiler/codegen/RuntimeStubsGenerator.scala
@@ -1,14 +1,16 @@
 package org.enso.compiler.codegen
 
+import org.enso.compiler.exception.CompilerError
 import org.enso.compiler.pass.analyse.BindingAnalysis
 import org.enso.interpreter.runtime.Module
+import org.enso.interpreter.runtime.builtin.Builtins
 import org.enso.interpreter.runtime.callable.atom.AtomConstructor
 
 /** Generates stubs of runtime representations of atom constructors, to allow
   * [[IrToTruffle the code generator]] to refer to constructors that are not
   * fully generated yet.
   */
-class RuntimeStubsGenerator() {
+class RuntimeStubsGenerator(builtins: Builtins) {
 
   /** Runs the stage on the given module.
     *
@@ -22,8 +24,17 @@ class RuntimeStubsGenerator() {
       "Non-parsed module used in stubs generator"
     )
     localBindings.constructors.foreach { tp =>
-      val constructor = new AtomConstructor(tp.name, scope)
-      scope.registerConstructor(constructor)
+      if (tp.builtinType) {
+        val builtinType = builtins.getBuiltinType(tp.name)
+        if (builtinType == null) {
+          throw new CompilerError("Unknown @BuiltinType " + tp.name)
+        }
+        scope.registerConstructor(builtinType)
+        builtinType.setShadowDefinitions(scope)
+      } else {
+        val constructor = new AtomConstructor(tp.name, scope)
+        scope.registerConstructor(constructor)
+      }
     }
   }
 }