+++ engine/runtime/src/main/scala/org/enso/compiler/data/BindingsMap.scala
@@ -735,8 +735,14 @@ object BindingsMap {
     * @param name the name of the constructor.
     * @param arity the number of fields in the constructor.
     * @param allFieldsDefaulted whether all fields provide a default value.
+    * @param builtinType true if constructor is annotated with @Builtin_Type, false otherwise.
     */
-  case class Cons(name: String, arity: Int, allFieldsDefaulted: Boolean)
+  case class Cons(
+    name: String,
+    arity: Int,
+    allFieldsDefaulted: Boolean,
+    builtinType: Boolean = false
+  )
 
   /** A representation of a sum type
     *