+++ engine/runtime/src/main/scala/org/enso/compiler/pass/analyse/BindingAnalysis.scala
@@ -11,7 +11,11 @@ import org.enso.compiler.pass.desugar.{
   FunctionBinding,
   GenerateMethodBodies
 }
-import org.enso.compiler.pass.resolve.{MethodDefinitions, Patterns}
+import org.enso.compiler.pass.resolve.{
+  MethodDefinitions,
+  ModuleAnnotations,
+  Patterns
+}
 
 import scala.annotation.unused
 
@@ -53,10 +57,15 @@ case object BindingAnalysis extends IRPass {
 
     val definedConstructors = ir.bindings.collect {
       case cons: IR.Module.Scope.Definition.Atom =>
+        // FIXME: move to a different pass
+        val isBuiltinType = cons
+          .getMetadata(ModuleAnnotations)
+          .exists(_.annotations.exists(_.name == "@Builtin_Type"))
         BindingsMap.Cons(
           cons.name.name,
           cons.arguments.length,
-          cons.arguments.forall(_.defaultValue.isDefined)
+          cons.arguments.forall(_.defaultValue.isDefined),
+          isBuiltinType
         )
     }
     val importedPolyglot = ir.imports.collect {