+++ engine/runtime/src/test/scala/org/enso/interpreter/test/semantic/LambdaShorthandArgsTest.scala
@@ -50,7 +50,7 @@ class LambdaShorthandArgsTest extends InterpreterTest {
 
     "work with mixfix functions" in {
       val code =
-        """from Standard.Builtins import all
+        """from Standard.Base.Data.Numbers import all
           |
           |Number.if_then_else = ~t -> ~f -> if this == 0 then t else f
           |
@@ -66,7 +66,7 @@ class LambdaShorthandArgsTest extends InterpreterTest {
 
     "work with case expressions" in {
       val code =
-        """from Standard.Builtins import all
+        """from Standard.Base.Data.List import all
           |
           |main =
           |    f = case _ of
@@ -142,9 +142,9 @@ class LambdaShorthandArgsTest extends InterpreterTest {
     "work properly when used with dot notation" in {
       val code =
         """
-          |import Standard.Builtins
+          |from Standard.Base.Data.Numbers import Number
           |
-          |Builtins.Number.f = this + 10
+          |Number.f = this + 10
           |
           |main =
           |    fun = _.f
@@ -157,8 +157,6 @@ class LambdaShorthandArgsTest extends InterpreterTest {
     "work properly when used inside the function of an application" in {
       val code =
         """
-          |import Standard.Builtins
-          |
           |main = (_ - 5) 0
           |""".stripMargin