+++ CHANGELOG.md
@@ -191,6 +191,7 @@
 - [Provide `tagValues` for function arguments in the language server][3422]
 - [Delay construction of Truffle nodes to speed initialization][3429]
 - [Frgaal compiler integration to allow for latest Java constructs][3421]
+- [Move Builtin Types and Methods definitions to stdlib][3363]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -206,6 +207,7 @@
 [3422]: https://github.com/enso-org/enso/pull/3422
 [3429]: https://github.com/enso-org/enso/pull/3429
 [3421]: https://github.com/enso-org/enso/pull/3421
+[3363]: https://github.com/enso-org/enso/pull/3363
 
 # Enso 2.0.0-alpha.18 (2021-10-12)